refactor(automations): remove K8s graph-alert cron, event-sourced path is sole path (ADR-034)
Every project has cut over to the event-sourced graph-alert path, so the
K8s cron is dead weight and the release_es_graph_triggers_firing flag is
no longer a meaningful choice. Removes both.

- Delete the /api/cron/triggers route + the whole pages/api/cron/triggers/
  handler tree (customGraphTrigger, actions/*, utils, cronGraphAlertDeps).
  Relocate the two still-used types (ActionParams, TriggerData) to
  app-layer/triggers/trigger.types.ts; the cron-only types die with it.
- Make the ES path unconditional: drop the flag gate from both
  graphTriggerEvaluation reactors and the heartbeat's project filter
  (the heartbeat now processes every project with graph triggers).
- Remove the release_es_graph_triggers_firing flag from the registry.
- Remove the alertTriggers Helm CronJob.
- Update the affected tests (delete the cron + dual-path suites; drop the
  flag mocks/dep from the reactor + heartbeat suites), ADR-034/040/044,
  the event-sourced-graph-triggers spec, and stale cron-referencing comments.

IMPORTANT: the ES path is worker-only (outbox drainer + heartbeat are
roleRunsWorkers-gated). A deployment MUST run the workers Deployment for
graph alerts to fire — there is no cron fallback anymore. Documented in
ADR-034 and the chart notes.

typecheck clean; 216 affected test files green.
