+++ dev/docs/adr/034-event-sourced-analytics-materialization.md
@@ -103,6 +103,7 @@ sums:    CostSum, NonBilledCostSum, DurationSum (root span),
 2. Both projections dual-tap **silently** (no read change). Reconstruct via **replay** (ADR-015) — slim fold re-applies idempotently; rollup is truncated and rebuilt.
 3. Repoint `getTimeseries` behind a Project flag; reconcile vs the live path (~1 week).
 4. Flip the flag; triggers move off the cron; keep the comparator as a tripwire.
+5. **Done (2026-07):** every project cut over, so the K8s cron graph-alert path (`/api/cron/triggers` + `pages/api/cron/triggers/`) and the `release_es_graph_triggers_firing` flag were **removed** — the event-sourced path is now unconditional. See the Implementation status note below.
 
 ## Keep / drop (from the prior attempt)
 
@@ -122,7 +123,7 @@ Shipped on branch `feat/event-sourced-graph-triggers` as a stack of phase-scoped
 - **Per-span source for the rollup is an app-side map projection**, not a ClickHouse materialized view. Cost is a real per-span column on `stored_spans` (Phase 0 migration `00037`), populated by the same `SpanCostService` the trace-summary fold uses; the rollup map projection reads those typed columns directly.
 - **Slim fold continuity** comes from `refoldOnStoreMiss` + a RedisCachedFoldStore wrapper (the slim row is lossy, no read-back) — see "Fold-state continuity" in the Decision. The **eval rollup** additionally enables the map-projection `dedupeByIdempotencyKey` option: eval terminal events are re-reported by design (deterministic ids, SDK retries), so without it every duplicate append double-counted the bucket — unlike the accepted per-span crash-retry noise.
 - **Read routing** lives in `src/server/app-layer/analytics/routing/route-table.ts` (`pickAnalyticsTable`); the union is `"trace_analytics_rollup" | "trace_analytics" | "trace_summaries" | "evaluation_analytics_rollup" | "evaluation_analytics" | "evaluation_runs"`. Per-project gating via the LangWatch (PostHog) feature flag `release_event_sourced_analytics_read`; optional tripwire via `release_event_sourced_analytics_read_tripwire`.
-- **Graph triggers off the cron**: real-time path is `.withOutbox("<aggregate>Analytics", "graphTriggerEvaluation", …)` on the trace AND eval pipelines, debounced per `(triggerId, projectId)` via `makeJobId` + 5 s ttl. Absence/resolve path is the 30 s outbox heartbeat (ADR-039) with a source-aware pre-filter against the right slim table per candidate; its no-data predicate is exactly "would a zero value breach?" for cron parity (a `count < 10` alert still fires on total silence). Cron coexists and skips flagged projects; per-project flag `release_es_graph_triggers_firing` flips a project onto the new path.
+- **Graph triggers off the cron**: real-time path is `.withOutbox("<aggregate>Analytics", "graphTriggerEvaluation", …)` on the trace AND eval pipelines, debounced per `(triggerId, projectId)` via `makeJobId` + 5 s ttl. Absence/resolve path is the 30 s outbox heartbeat (ADR-039) with a source-aware pre-filter against the right slim table per candidate; its no-data predicate is exactly "would a zero value breach?" for parity with the old cron (a `count < 10` alert still fires on total silence). **Migration complete (2026-07):** once every project had cut over, the K8s cron path and the `release_es_graph_triggers_firing` flag were removed — the event-sourced path is now the sole, unconditional graph-alert path. This runs on the worker role only (outbox drainer + heartbeat are `roleRunsWorkers`-gated), so a deployment MUST run the workers Deployment for graph alerts to fire.
 - **Scenarios / experiments / suites** were originally built write-side only (Phase 7) and then **pulled back** during the 2026-07-07 stack review — see the Scope note under Shared fleet. Re-add each aggregate when a real consumer names its columns.
 - **UI**: dashboard "Add alert" + the `/analytics/custom` page repoint at the automations drawer pre-filled with `prefilledGraphId` + `prefilledSeriesName` (Phase 5.2 + 8). The legacy `AlertDrawer.tsx` was verified unreachable and **deleted** in the housekeeping pass at the end of the stack. Automations drawer disables non-notify action cards when source = customGraph (server-side validation rejects them too).
 - **Alert template data is graph-shaped, not trace-shaped**: `GraphAlertTemplateContext` carries the metric's recent numeric `history` (the buckets the threshold read — no extra query), a prebuilt `sparkline`, `previousValue` (preceding window, same aggregation), and a `graph.url` deep link pinned to the incident window. Preview, variable autocomplete, and test-fire render the same alert shape (`buildExampleGraphAlertTemplateContext` + `ALERT_TRIGGER_DEFAULTS`) so what the author sees matches what dispatch sends. Rendering the history as a PNG chart attachment was considered and deferred — it needs image hosting for Slack and CID inlining for email; `history` keeps that door open.
