--- langwatch/src/pages/api/cron/triggers/__tests__/customGraphTrigger.test.ts
@@ -1,853 +1,0 @@
-import { type Project, type Trigger, TriggerAction } from "@prisma/client";
-import { beforeEach, describe, expect, it, vi } from "vitest";
-import type { TimeseriesBucket } from "~/server/analytics/types";
-import { sumMetricAcrossGroups } from "~/server/app-layer/analytics/series-points";
-import { processCustomGraphTrigger } from "../customGraphTrigger";
-
-const { mockGetTimeseries } = vi.hoisted(() => ({
-  mockGetTimeseries: vi.fn(),
-}));
-vi.mock("~/server/app-layer/analytics/analytics.service", () => ({
-  getAnalyticsService: () => ({
-    getTimeseries: mockGetTimeseries,
-  }),
-}));
-
-vi.mock("~/server/db", () => ({
-  prisma: {
-    customGraph: {
-      findUnique: vi.fn(),
-    },
-    triggerSent: {
-      findFirst: vi.fn(),
-      update: vi.fn(),
-    },
-  },
-}));
-
-vi.mock("../actions/sendEmail", () => ({
-  handleSendEmail: vi.fn(),
-}));
-
-vi.mock("../actions/sendSlackMessage", () => ({
-  handleSendSlackMessage: vi.fn(),
-}));
-
-vi.mock("../utils", () => ({
-  updateAlert: vi.fn(),
-  checkThreshold: vi.fn(),
-  addTriggersSent: vi.fn(),
-}));
-
-vi.mock("~/utils/posthogErrorCapture", () => ({
-  captureException: vi.fn(),
-  toError: vi.fn((e) => e instanceof Error ? e : new Error(String(e))),
-}));
-
-import { prisma } from "~/server/db";
-import { captureException } from "~/utils/posthogErrorCapture";
-import { handleSendEmail } from "../actions/sendEmail";
-import { handleSendSlackMessage } from "../actions/sendSlackMessage";
-import { addTriggersSent, checkThreshold, updateAlert } from "../utils";
-
-describe("processCustomGraphTrigger", () => {
-  const mockProjects: Project[] = [
-    { id: "project-1", slug: "test-project" } as Project,
-  ];
-
-  const makeFiringTrigger = (action: TriggerAction) =>
-    ({
-      id: "trigger-1",
-      name: "Test Alert",
-      projectId: "project-1",
-      customGraphId: "graph-1",
-      action,
-      actionParams: {
-        threshold: 10,
-        operator: "gt",
-        timePeriod: 60,
-        members: [],
-        seriesName: "0/count/count",
-      },
-    }) as unknown as Trigger;
-
-  const mockFiringGraph = () => {
-    vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-      id: "graph-1",
-      name: "Test Graph",
-      graph: {
-        series: [{ name: "metric1", metric: "count", aggregation: "count" }],
-      },
-      filters: {},
-    } as any);
-    mockGetTimeseries.mockResolvedValue({
-      currentPeriod: [{ "0/count/count": 15 }],
-      previousPeriod: [],
-    } as any);
-    vi.mocked(checkThreshold).mockReturnValue(true);
-  };
-
-  beforeEach(() => {
-    vi.clearAllMocks();
-    // Default mock: no unresolved trigger (new alert will be sent)
-    vi.mocked(prisma.triggerSent.findFirst).mockResolvedValue(null);
-    // Default: dispatch succeeds — the handlers report the fire as consumed.
-    vi.mocked(handleSendEmail).mockResolvedValue({ didSend: true });
-    vi.mocked(handleSendSlackMessage).mockResolvedValue({ didSend: true });
-  });
-
-  describe("when trigger has no customGraphId", () => {
-    it("returns error status with message", async () => {
-      const trigger = {
-        id: "trigger-1",
-        projectId: "project-1",
-        customGraphId: null,
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "0/metadata.trace_id/cardinality",
-        },
-      } as unknown as Trigger;
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result).toEqual({
-        triggerId: "trigger-1",
-        status: "error",
-        message: "No customGraphId found",
-      });
-    });
-  });
-
-  describe("when custom graph is not found", () => {
-    it("returns error status with message", async () => {
-      const trigger = {
-        id: "trigger-1",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "0/metadata.trace_id/cardinality",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue(null);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result).toEqual({
-        triggerId: "trigger-1",
-        status: "error",
-        message: "Graph not found",
-      });
-    });
-  });
-
-  describe("when graph has no series", () => {
-    it("returns error status with message", async () => {
-      const trigger = {
-        id: "trigger-1",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "0/metadata.trace_id/cardinality",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: { series: [] },
-        filters: {},
-      } as any);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result).toEqual({
-        triggerId: "trigger-1",
-        status: "error",
-        message: "No series found in graph",
-      });
-    });
-  });
-
-  describe("when seriesName is missing", () => {
-    it("returns error status with message", async () => {
-      const trigger = {
-        id: "trigger-1",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: {
-          series: [{ name: "metric1", metric: "count", aggregation: "count" }],
-        },
-        filters: {},
-      } as any);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result).toEqual({
-        triggerId: "trigger-1",
-        status: "error",
-        message: "seriesName is required in ActionParams",
-      });
-    });
-  });
-
-  describe("when seriesName has invalid index", () => {
-    it("returns error status when index is out of bounds", async () => {
-      const trigger = {
-        id: "trigger-1",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "5/count/count",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: {
-          series: [{ name: "metric1", metric: "count", aggregation: "count" }],
-        },
-        filters: {},
-      } as any);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result).toEqual({
-        triggerId: "trigger-1",
-        status: "error",
-        message: "Series index 5 not found in graph (has 1 series)",
-      });
-    });
-  });
-
-  describe("when monitoring second series in multi-series graph", () => {
-    it("monitors the correct series based on index", async () => {
-      const trigger = {
-        id: "trigger-1",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        action: TriggerAction.SEND_EMAIL,
-        actionParams: {
-          threshold: 50,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "1/errors/count",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: {
-          series: [
-            { name: "successes", metric: "success", aggregation: "count" },
-            { name: "errors", metric: "errors", aggregation: "count" },
-          ],
-        },
-        filters: {},
-      } as any);
-
-      vi.mocked(prisma.triggerSent.findFirst).mockResolvedValue(null);
-
-      // seriesName "1/errors/count" selects series INDEX 1 (the "errors"
-      // series); the evaluator queries it as a single-series input, so the
-      // analytics result bucket is keyed "0/errors/count" (query index 0) —
-      // NOT the stored "1/errors/count". The evaluator derives that lookup
-      // key via buildSeriesName, matching what the analytics service returns.
-      mockGetTimeseries.mockResolvedValue({
-        currentPeriod: [{ "0/errors/count": 60 }, { "0/errors/count": 70 }],
-        previousPeriod: [],
-      } as any);
-      vi.mocked(checkThreshold).mockReturnValue(true);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result).toEqual({
-        triggerId: "trigger-1",
-        status: "triggered",
-        value: 130,
-        threshold: 50,
-        operator: "gt",
-      });
-    });
-  });
-
-  describe("when threshold condition is met", () => {
-    it("returns triggered status and updates trigger", async () => {
-      const trigger = {
-        id: "trigger-1",
-        name: "Test Alert",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        action: TriggerAction.SEND_EMAIL,
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          members: [],
-          seriesName: "0/count/count",
-        },
-        message: "Custom message",
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: {
-          series: [
-            {
-              name: "metric1",
-              metric: "count",
-              aggregation: "count",
-            },
-          ],
-          timeScale: 60,
-        },
-        filters: {},
-      } as any);
-
-      mockGetTimeseries.mockResolvedValue({
-        currentPeriod: [{ "0/count/count": 15 }, { "0/count/count": 20 }],
-        previousPeriod: [],
-      } as any);
-
-      vi.mocked(checkThreshold).mockReturnValue(true);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result).toEqual({
-        triggerId: "trigger-1",
-        status: "triggered",
-        value: 35,
-        threshold: 10,
-        operator: "gt",
-      });
-
-      expect(handleSendEmail).toHaveBeenCalled();
-      expect(updateAlert).toHaveBeenCalledWith(
-        "trigger-1",
-        expect.any(Number),
-        "project-1",
-      );
-    });
-  });
-
-  describe("when threshold condition is not met", () => {
-    it("returns not_triggered status and updates trigger", async () => {
-      const trigger = {
-        id: "trigger-1",
-        name: "Test Alert",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        action: TriggerAction.SEND_EMAIL,
-        actionParams: {
-          threshold: 100,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "0/count/count",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: {
-          series: [
-            {
-              name: "metric1",
-              metric: "count",
-              aggregation: "count",
-            },
-          ],
-        },
-        filters: {},
-      } as any);
-
-      mockGetTimeseries.mockResolvedValue({
-        currentPeriod: [{ "0/count/count": 5 }],
-        previousPeriod: [],
-      } as any);
-      vi.mocked(checkThreshold).mockReturnValue(false);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result).toEqual({
-        triggerId: "trigger-1",
-        status: "not_triggered",
-        value: 5,
-        threshold: 100,
-        operator: "gt",
-      });
-
-      expect(handleSendEmail).not.toHaveBeenCalled();
-      expect(updateAlert).toHaveBeenCalled();
-    });
-  });
-
-  describe("when action is SEND_SLACK_MESSAGE", () => {
-    it("calls handleSendSlackMessage with graph alert context", async () => {
-      const trigger = {
-        id: "trigger-1",
-        name: "Test Alert",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        action: TriggerAction.SEND_SLACK_MESSAGE,
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          slackWebhook: "https://hooks.slack.com/test",
-          seriesName: "0/count/count",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: {
-          series: [{ name: "metric1", metric: "count", aggregation: "count" }],
-        },
-        filters: {},
-      } as any);
-
-      mockGetTimeseries.mockResolvedValue({
-        currentPeriod: [{ "0/count/count": 15 }],
-        previousPeriod: [],
-      } as any);
-      vi.mocked(checkThreshold).mockReturnValue(true);
-
-      await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(handleSendSlackMessage).toHaveBeenCalled();
-    });
-  });
-
-  describe("when the dispatch reports the fire as consumed (didSend true)", () => {
-    it("records the incident via addTriggersSent", async () => {
-      const trigger = makeFiringTrigger(TriggerAction.SEND_EMAIL);
-      mockFiringGraph();
-
-      await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(addTriggersSent).toHaveBeenCalledWith(
-        "trigger-1",
-        expect.any(Array),
-      );
-    });
-  });
-
-  // Regression: the handlers used to swallow every dispatch error into void,
-  // and the caller recorded the incident unconditionally — an undelivered
-  // alert read as "currently firing" and suppressed all future notifications.
-  describe("when the email dispatch fails (didSend false)", () => {
-    it("does not record the incident, so the next tick retries", async () => {
-      vi.mocked(handleSendEmail).mockResolvedValue({ didSend: false });
-      const trigger = makeFiringTrigger(TriggerAction.SEND_EMAIL);
-      mockFiringGraph();
-
-      await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(handleSendEmail).toHaveBeenCalled();
-      expect(addTriggersSent).not.toHaveBeenCalled();
-    });
-  });
-
-  describe("when the Slack dispatch fails (didSend false)", () => {
-    it("does not record the incident, so the next tick retries", async () => {
-      vi.mocked(handleSendSlackMessage).mockResolvedValue({ didSend: false });
-      const trigger = makeFiringTrigger(TriggerAction.SEND_SLACK_MESSAGE);
-      mockFiringGraph();
-
-      await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(handleSendSlackMessage).toHaveBeenCalled();
-      expect(addTriggersSent).not.toHaveBeenCalled();
-    });
-  });
-
-  // Regression: the resolve used to write only `resolvedAt`, leaving
-  // `openIncidentKey` held — so the event-sourced path's next
-  // `claimOpenForGraphAlert` INSERT hit the unique key (P2002) forever and the
-  // alert could never fire again. Mirrors `markResolvedById` in the trigger
-  // prisma repository.
-  describe("when a firing alert resolves", () => {
-    it("clears openIncidentKey alongside resolvedAt", async () => {
-      vi.mocked(prisma.triggerSent.findFirst).mockResolvedValue({
-        id: "sent-1",
-      } as any);
-      const trigger = makeFiringTrigger(TriggerAction.SEND_EMAIL);
-      mockFiringGraph();
-      // Condition no longer met: the open incident must resolve.
-      vi.mocked(checkThreshold).mockReturnValue(false);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result.status).toBe("not_triggered");
-      expect(prisma.triggerSent.update).toHaveBeenCalledWith({
-        where: { id: "sent-1", projectId: "project-1" },
-        data: { resolvedAt: expect.any(Date), openIncidentKey: null },
-      });
-    });
-  });
-
-  describe("when graph uses groupBy", () => {
-    it("sums grouped metric values across groups per time bucket", async () => {
-      const trigger = {
-        id: "trigger-1",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        action: TriggerAction.SEND_EMAIL,
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "0/metadata.trace_id/cardinality",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: {
-          series: [
-            {
-              name: "traces",
-              metric: "metadata.trace_id",
-              aggregation: "cardinality",
-            },
-          ],
-          groupBy: "sentiment.thumbs_up_down",
-        },
-        filters: {},
-      } as any);
-
-      mockGetTimeseries.mockResolvedValue({
-        currentPeriod: [
-          {
-            date: "2024-01-01",
-            "sentiment.thumbs_up_down": {
-              "Thumbs Up": {
-                "0/metadata.trace_id/cardinality": 10,
-              },
-              "Thumbs Down": {
-                "0/metadata.trace_id/cardinality": 5,
-              },
-            },
-          },
-        ],
-        previousPeriod: [],
-      });
-
-      vi.mocked(checkThreshold).mockReturnValue(true);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      // 10 + 5 = 15, summed across groups
-      expect(result.value).toBe(15);
-    });
-
-    it("evaluates to 0 when no group contains the monitored metric", async () => {
-      const trigger = {
-        id: "trigger-1",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        action: TriggerAction.SEND_EMAIL,
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "0/metadata.trace_id/cardinality",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: {
-          series: [
-            {
-              name: "traces",
-              metric: "metadata.trace_id",
-              aggregation: "cardinality",
-            },
-          ],
-          groupBy: "sentiment.thumbs_up_down",
-        },
-        filters: {},
-      } as any);
-
-      mockGetTimeseries.mockResolvedValue({
-        currentPeriod: [
-          {
-            date: "2024-01-01",
-            "sentiment.thumbs_up_down": {
-              "Thumbs Up": { "other/metric": 99 },
-            },
-          },
-        ],
-        previousPeriod: [],
-      });
-
-      vi.mocked(checkThreshold).mockReturnValue(false);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result.value).toBe(0);
-    });
-  });
-
-  describe("when timeseries calculation uses average aggregation", () => {
-    it("calculates current value as average of all values", async () => {
-      const trigger = {
-        id: "trigger-1",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        action: TriggerAction.SEND_EMAIL,
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "0/count/avg",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: {
-          series: [
-            {
-              name: "metric1",
-              metric: "count",
-              aggregation: "avg",
-            },
-          ],
-        },
-        filters: {},
-      } as any);
-
-      mockGetTimeseries.mockResolvedValue({
-        currentPeriod: [
-          { "0/count/avg": 10 },
-          { "0/count/avg": 20 },
-          { "0/count/avg": 30 },
-        ],
-        previousPeriod: [],
-      } as any);
-
-      vi.mocked(checkThreshold).mockReturnValue(false);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result.value).toBe(20); // (10 + 20 + 30) / 3
-    });
-  });
-
-  describe("when timeseries returns no data", () => {
-    it("sets current value to 0", async () => {
-      const trigger = {
-        id: "trigger-1",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        action: TriggerAction.SEND_EMAIL,
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "0/count/count",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: {
-          series: [{ name: "metric1", metric: "count", aggregation: "count" }],
-        },
-        filters: {},
-      } as any);
-
-      mockGetTimeseries.mockResolvedValue({
-        currentPeriod: [],
-        previousPeriod: [],
-      } as any);
-      vi.mocked(checkThreshold).mockReturnValue(false);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result.value).toBe(0);
-    });
-  });
-
-  describe("when project is not found", () => {
-    it("returns error status with message", async () => {
-      const trigger = {
-        id: "trigger-1",
-        projectId: "unknown-project",
-        customGraphId: "graph-1",
-        action: TriggerAction.SEND_EMAIL,
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "0/count/count",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockResolvedValue({
-        id: "graph-1",
-        name: "Test Graph",
-        graph: {
-          series: [{ name: "metric1", metric: "count", aggregation: "count" }],
-        },
-        filters: {},
-      } as any);
-
-      mockGetTimeseries.mockResolvedValue({
-        currentPeriod: [{ "0/count/count": 15 }],
-        previousPeriod: [],
-      } as any);
-      vi.mocked(checkThreshold).mockReturnValue(true);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result).toEqual({
-        triggerId: "trigger-1",
-        status: "error",
-        message: "Project not found",
-      });
-    });
-  });
-
-  describe("when an exception occurs", () => {
-    it("captures exception and returns error status", async () => {
-      const error = new Error("Database error");
-      const trigger = {
-        id: "trigger-1",
-        projectId: "project-1",
-        customGraphId: "graph-1",
-        actionParams: {
-          threshold: 10,
-          operator: "gt",
-          timePeriod: 60,
-          seriesName: "0/count/count",
-        },
-      } as unknown as Trigger;
-
-      vi.mocked(prisma.customGraph.findUnique).mockRejectedValue(error);
-
-      const result = await processCustomGraphTrigger(trigger, mockProjects);
-
-      expect(result).toEqual({
-        triggerId: "trigger-1",
-        status: "error",
-        message: "Database error",
-      });
-
-      expect(captureException).toHaveBeenCalledWith(error, {
-        extra: {
-          triggerId: "trigger-1",
-          projectId: "project-1",
-          type: "customGraphAlert",
-        },
-      });
-    });
-  });
-});
-
-describe("sumMetricAcrossGroups()", () => {
-  const seriesKey = "0/metadata.trace_id/cardinality";
-
-  describe("when entry has grouped data with the metric", () => {
-    it("sums the metric across all groups", () => {
-      const entry: TimeseriesBucket = {
-        date: "2024-01-01",
-        "sentiment.thumbs_up_down": {
-          "Thumbs Up": { [seriesKey]: 10 },
-          "Thumbs Down": { [seriesKey]: 5 },
-        },
-      };
-
-      expect(
-        sumMetricAcrossGroups(entry, "sentiment.thumbs_up_down", seriesKey),
-      ).toBe(15);
-    });
-  });
-
-  describe("when some groups lack the metric", () => {
-    it("sums only the groups that have it", () => {
-      const entry: TimeseriesBucket = {
-        date: "2024-01-01",
-        "metadata.model": {
-          "gpt-4": { [seriesKey]: 8 },
-          "claude": { "other/metric": 99 },
-          "gemini": { [seriesKey]: 2 },
-        },
-      };
-
-      expect(
-        sumMetricAcrossGroups(entry, "metadata.model", seriesKey),
-      ).toBe(10);
-    });
-  });
-
-  describe("when groupBy key is missing from entry", () => {
-    it("returns undefined", () => {
-      const entry: TimeseriesBucket = { date: "2024-01-01" };
-
-      expect(
-        sumMetricAcrossGroups(entry, "sentiment.thumbs_up_down", seriesKey),
-      ).toBeUndefined();
-    });
-  });
-
-  describe("when no group contains the metric", () => {
-    it("returns undefined", () => {
-      const entry: TimeseriesBucket = {
-        date: "2024-01-01",
-        "sentiment.thumbs_up_down": {
-          "Thumbs Up": { "other/metric": 10 },
-        },
-      };
-
-      expect(
-        sumMetricAcrossGroups(entry, "sentiment.thumbs_up_down", seriesKey),
-      ).toBeUndefined();
-    });
-  });
-
-  describe("when groupBy value is not an object", () => {
-    it("returns undefined", () => {
-      const entry: TimeseriesBucket = {
-        date: "2024-01-01",
-        "metadata.model": 42,
-      };
-
-      expect(
-        sumMetricAcrossGroups(entry, "metadata.model", seriesKey),
-      ).toBeUndefined();
-    });
-  });
-});
