+++ langwatch/src/server/app-layer/triggers/__tests__/graph-trigger-heartbeat.unit.test.ts
@@ -2,7 +2,6 @@ import type { ClickHouseClient } from "@clickhouse/client";
 import { TriggerAction, TriggerKind } from "@prisma/client";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import type { GraphEvalStagePayload } from "~/server/event-sourcing/outbox/payload";
-import type { FeatureFlagServiceInterface } from "~/server/featureFlag/types";
 import {
   decideGraphTriggerHeartbeat,
   defaultCandidateSources,
@@ -60,12 +59,6 @@ function makeTriggersService(
     updateLastRunAt: vi.fn(),
     invalidate: vi.fn(),
   } as unknown as TriggerService;
-}
-
-function makeFlagsAllOn(): FeatureFlagServiceInterface {
-  return {
-    isEnabled: vi.fn(async () => true),
-  };
 }
 
 function makeSources(overrides: {
@@ -127,31 +120,19 @@ describe("decideGraphTriggerHeartbeat", () => {
     chStub = makeClickHouseStub({});
   });
 
-  describe("given no flagged projects", () => {
+  describe("given no candidate projects", () => {
     it("returns no enqueues", async () => {
-      const triggers = makeTriggersService({
-        [PROJECT_A]: [
-          makeTrigger(TRIGGER_NO_DATA, PROJECT_A, {
-            threshold: 1,
-            operator: "lt",
-            timePeriod: 5,
-          }),
-        ],
-      });
-      const flags: FeatureFlagServiceInterface = {
-        isEnabled: vi.fn(async () => false),
-      };
+      const triggers = makeTriggersService({});
       deps = {
         triggers,
         prisma: prismaStub as unknown as GraphTriggerHeartbeatDeps["prisma"],
         resolveClickHouseClient: async () => chStub.client,
-        featureFlagService: flags,
         lookupTriggerSource: async () => "trace" as const,
       };
 
       const result = await decideGraphTriggerHeartbeat({
         deps,
-        sources: makeSources({ graphProjects: [PROJECT_A] }),
+        sources: makeSources({ graphProjects: [] }),
         now,
       });
 
@@ -174,7 +155,6 @@ describe("decideGraphTriggerHeartbeat", () => {
         triggers,
         prisma: prismaStub as unknown as GraphTriggerHeartbeatDeps["prisma"],
         resolveClickHouseClient: async () => chStub.client,
-        featureFlagService: makeFlagsAllOn(),
         lookupTriggerSource: async () => "trace" as const,
       };
 
@@ -204,7 +184,6 @@ describe("decideGraphTriggerHeartbeat", () => {
         triggers,
         prisma: prismaStub as unknown as GraphTriggerHeartbeatDeps["prisma"],
         resolveClickHouseClient: async () => chStub.client,
-        featureFlagService: makeFlagsAllOn(),
         lookupTriggerSource: async () => "trace" as const,
       };
 
@@ -241,7 +220,6 @@ describe("decideGraphTriggerHeartbeat", () => {
         triggers,
         prisma: prismaStub as unknown as GraphTriggerHeartbeatDeps["prisma"],
         resolveClickHouseClient: async () => chStub.client,
-        featureFlagService: makeFlagsAllOn(),
         lookupTriggerSource: async () => "trace" as const,
       };
 
@@ -273,7 +251,6 @@ describe("decideGraphTriggerHeartbeat", () => {
         triggers,
         prisma: prismaStub as unknown as GraphTriggerHeartbeatDeps["prisma"],
         resolveClickHouseClient: async () => chStub.client,
-        featureFlagService: makeFlagsAllOn(),
         lookupTriggerSource: async () => "trace" as const,
       };
 
@@ -318,7 +295,6 @@ describe("decideGraphTriggerHeartbeat", () => {
         triggers,
         prisma: prismaStub as unknown as GraphTriggerHeartbeatDeps["prisma"],
         resolveClickHouseClient: async () => chStub.client,
-        featureFlagService: makeFlagsAllOn(),
         lookupTriggerSource: async () => "trace" as const,
       };
