+++ langwatch/src/server/mailer/__tests__/triggerEmail.unit.test.tsx
@@ -1,6 +1,6 @@
 import { AlertType } from "@prisma/client";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import type { TriggerData } from "~/pages/api/cron/triggers/types";
+import type { TriggerData } from "~/server/app-layer/triggers/trigger.types";
 import { DispatchError } from "~/server/event-sourcing/outbox/dispatchError";
 import type { Trace } from "~/server/tracer/types";
