+++ langwatch/src/server/mailer/triggerEmail.tsx
@@ -8,7 +8,7 @@ import { Img } from "@react-email/img";
 import { render } from "@react-email/render";
 import { createHash } from "crypto";
 import { EMAIL_RX } from "~/automations/providers/definitions/email/shared";
-import type { TriggerData } from "~/pages/api/cron/triggers/types";
+import type { TriggerData } from "~/server/app-layer/triggers/trigger.types";
 import { toDispatchError } from "~/server/event-sourcing/outbox/dispatchError";
 import { env } from "../../env.mjs";
 import { computeDefaultFrom, sendEmail } from "./emailSender";
