+++ langwatch/src/server/routes/__tests__/internal-routes-auth.integration.test.ts
@@ -21,7 +21,6 @@ const CRON_ROUTES: { method: "GET" | "POST"; path: string }[] = [
   { method: "POST", path: "/api/cron/schedule_topic_clustering" },
   { method: "GET", path: "/api/cron/schedule_topic_clustering" },
   { method: "GET", path: "/api/cron/trace_analytics" },
-  { method: "GET", path: "/api/cron/triggers" },
   { method: "POST", path: "/api/cron/seed_demo" },
   { method: "GET", path: "/api/cron/seed_demo" },
 ];
